+++ apollo-client/src/test/java/com/ctrip/framework/apollo/spring/JavaConfigPlaceholderTest.java
@@ -17,6 +17,7 @@
 package com.ctrip.framework.apollo.spring;
 
 import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertNotNull;
 import static org.mockito.ArgumentMatchers.anyString;
 import static org.mockito.ArgumentMatchers.eq;
 import static org.mockito.Mockito.mock;
@@ -27,7 +28,10 @@ import com.ctrip.framework.apollo.PropertiesCompatibleConfigFile;
 import com.ctrip.framework.apollo.core.ConfigConsts;
 import com.ctrip.framework.apollo.spring.annotation.ApolloJsonValue;
 import com.ctrip.framework.apollo.spring.annotation.EnableApolloConfig;
+import java.text.SimpleDateFormat;
+import java.util.Date;
 import java.util.List;
+import java.util.Locale;
 import java.util.Properties;
 import org.junit.Test;
 import org.mockito.Mockito;
@@ -53,6 +57,9 @@ public class JavaConfigPlaceholderTest extends AbstractSpringIntegrationTest {
   private static final String FX_APOLLO_NAMESPACE = "FX.apollo";
   private static final String JSON_PROPERTY = "jsonProperty";
   private static final String OTHER_JSON_PROPERTY = "otherJsonProperty";
+  private static final String DATE_FORMAT_JSON_PROPERTY1 = "jsonDateProperty1";
+  private static final String DATE_FORMAT_JSON_PROPERTY2 = "jsonDateProperty2";
+  private static final String DATE_FORMAT_JSON_PROPERTY3 = "jsonDateProperty3";
 
   @Test
   public void testPropertySourceWithNoNamespace() throws Exception {
@@ -385,6 +392,30 @@ public class JavaConfigPlaceholderTest extends AbstractSpringIntegrationTest {
     assertEquals(20, testJsonPropertyBean.getOtherJsonBeanList().get(1).getB());
   }
 
+  @Test
+  public void testApolloDateJsonValue() {
+    String dateFormatJson1 = "{\"startTime\":\"2024-01-20\",\"endTime\":\"2024-01-20\"}";
+    String dateFormatJson2 = "{\"startTime\":\"2024-01-20T16:51:48\",\"endTime\":\"2024-01-20T16:51:48\"}";
+    String dateFormatJson3 = "{\"startTime\":\"2024/01/20\",\"endTime\":\"2024/01/20\"}";
+
+    String someDateFormat = "yyyy-MM-dd HH:mm:ss.SSS";
+    SimpleDateFormat simpleDateFormat = new SimpleDateFormat(someDateFormat, Locale.US);
+
+    Config config = mock(Config.class);
+    when(config.getProperty(eq(DATE_FORMAT_JSON_PROPERTY1), Mockito.nullable(String.class))).thenReturn(dateFormatJson1);
+    when(config.getProperty(eq(DATE_FORMAT_JSON_PROPERTY2), Mockito.nullable(String.class))).thenReturn(dateFormatJson2);
+    when(config.getProperty(eq(DATE_FORMAT_JSON_PROPERTY3), Mockito.nullable(String.class))).thenReturn(dateFormatJson3);
+    mockConfig(ConfigConsts.NAMESPACE_APPLICATION, config);
+
+    AnnotationConfigApplicationContext context = new AnnotationConfigApplicationContext(
+        AppConfig12.class);
+
+    TestJsonDatePropertyBean datePropertyBean = context.getBean(TestJsonDatePropertyBean.class);
+    assertEquals("2024-01-20 00:00:00.000", simpleDateFormat.format(datePropertyBean.getPattern1().getStartTime()));
+    assertEquals("2024-01-20 16:51:48.000", simpleDateFormat.format(datePropertyBean.getPattern2().getStartTime()));
+    assertEquals("2024-01-20 00:00:00.000", simpleDateFormat.format(datePropertyBean.getPattern3().getStartTime()));
+  }
+
   @Test(expected = BeanCreationException.class)
   public void testApolloJsonValueWithInvalidJson() throws Exception {
     String someInvalidJson = "someInvalidJson";
@@ -521,6 +552,15 @@ public class JavaConfigPlaceholderTest extends AbstractSpringIntegrationTest {
     @Bean
     TestJavaConfigBean testJavaConfigBean() {
       return new TestJavaConfigBean();
+    }
+  }
+
+  @Configuration
+  @EnableApolloConfig
+  static class AppConfig12 {
+    @Bean
+    TestJsonDatePropertyBean testJsonDatePropertyBean() {
+      return new TestJsonDatePropertyBean();
     }
   }
 
@@ -670,6 +710,59 @@ public class JavaConfigPlaceholderTest extends AbstractSpringIntegrationTest {
       int result = a != null ? a.hashCode() : 0;
       result = 31 * result + b;
       return result;
+    }
+  }
+
+  static class TestJsonDatePropertyBean {
+
+    @ApolloJsonValue(value = "${jsonDateProperty1}", datePattern = "yyyy-MM-dd")
+    private JsonDateBean pattern1;
+
+    @ApolloJsonValue(value = "${jsonDateProperty2}", datePattern = "yyyy-MM-dd'T'HH:mm:ss")
+    private JsonDateBean pattern2;
+
+    @ApolloJsonValue(value = "${jsonDateProperty3}", datePattern = "yyy/MM/dd")
+    private JsonDateBean pattern3;
+
+    public JsonDateBean getPattern1() {
+      return pattern1;
+    }
+
+    public JsonDateBean getPattern2() {
+      return pattern2;
+    }
+
+    public JsonDateBean getPattern3() {
+      return pattern3;
+    }
+  }
+
+  static class JsonDateBean {
+    private Date startTime;
+    private Date endTime;
+
+    public Date getStartTime() {
+      return startTime;
+    }
+
+    public void setStartTime(Date startTime) {
+      this.startTime = startTime;
+    }
+
+    public Date getEndTime() {
+      return endTime;
+    }
+
+    public void setEndTime(Date endTime) {
+      this.endTime = endTime;
+    }
+
+    @Override
+    public String toString() {
+      return "JsonDateBean{" +
+          "startTime=" + startTime +
+          ", endTime=" + endTime +
+          '}';
     }
   }
 }